Removed a few more 64 bits windows warnings 500+ to go.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8527 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/atermpp/include/mcrl2/atermpp/indexed_set.h
@@ -66,7 +66,7 @@ namespace atermpp
       std::pair<long, bool> put(aterm elem)
       {
         ATbool b;
-        long l = ATindexedSetPut(m_set.get(), elem, &b);
+        size_t l = ATindexedSetPut(m_set.get(), elem, &b);
         return std::make_pair(l, b == ATtrue);
       }
